Align article lookup numeric types with stored procedure floats

+++ src/ERPAccounting.Application/DTOs/ComboDtos.cs
@@ -59,11 +59,11 @@ public record ArticleComboDto(
     string NazivArtikla,
     string? JedinicaMere,
     string? IdPoreskaStopa,
-    decimal ProcenatPoreza,
-    decimal Akciza,
-    decimal KoeficijentKolicine,
+    double ProcenatPoreza,
+    double Akciza,
+    double KoeficijentKolicine,
     bool ImaLot,
-    decimal? OtkupnaCena,
+    double? OtkupnaCena,
     bool PoljoprivredniProizvod
 );
